+++ acme/etc/RequestUtils.py
@@ -379,3 +379,32 @@ def filterAttributes(dct:JSON, attributesToInclude:list[str]) -> JSON:
 			 for k, v in dct.items() 
 			 if k in attributesToInclude }
 			 
+
+
+def curlFromRequest(request:JSON) -> str:
+	"""	Create a cURL command from a request.
+	
+		Args:
+			request: The request to create the cURL command from.
+		
+		Return:
+			A cURL command.
+	"""
+	from ..runtime import CSE
+
+	if not request:
+		return 'No request available'
+
+	curl = f"""\
+curl -X {[None, 'POST', 'GET', 'PUT', 'DELETE', 'POST' ][request['op']]} '{CSE.httpServer.serverAddress}{CSE.httpServer.rootPath}/{request['to']}' \\
+  -H 'X-M2M-Origin: {request['fr']}' \\
+  -H 'X-M2M-RVI: {request['rvi']}' \\
+  -H 'X-M2M-RI: {request['rqi']}'"""
+	
+	if (ot := request.get('ot')):
+		curl += f" \\\n  -H 'X-M2M-OT: {ot}'"
+	if (pc := request.get('pc')):
+		curl += f" \\\n  -H 'Content-Type: {request['csz']}'"
+		curl += f" \\\n  -d '{json.dumps(pc)}'"
+	
+	return curl