+++ acme/textui/ACMEContainerResourceServices.py
@@ -10,6 +10,8 @@
 from __future__ import annotations
 from typing import cast
 
+import pyperclip
+
 from textual import on
 from textual.app import ComposeResult
 from textual.containers import Container, Horizontal, VerticalScroll, Vertical
@@ -48,7 +50,7 @@ class ACMEContainerResourceServices(Container):
 			# Export resource
 			with (v := Vertical(id = 'services-export-resource')):
 				v.border_title = 'Export Resource'
-				yield Label('Export the resource to a file in the *./tmp* directory as a *curl* command.', classes='label')
+				yield Label('Export the resource to a file in the [i]./tmp[/i] directory as a [i]curl[/i] command.', classes='label')
 				with Container(classes='service-command-area'):
 					with Horizontal(classes = 'services-export-controls'):
 						yield Button('Export', variant = 'primary', id = 'services-export-resource-button', classes = 'button')
@@ -59,9 +61,11 @@ class ACMEContainerResourceServices(Container):
 			# Export Instances
 			with (v := Vertical(id = 'services-export-instances')):
 				v.border_title = 'Export Instances'
-				yield Label('Export the instances of the container resource to a CSV file in the *./tmp* directory.', classes='label')
+				yield Label('Export the instances of the container resource to a [i]CSV[/i] file in the [i]./tmp[/i] directory or to the clipboard.', classes='label')
 				with Container(classes='service-command-area'):
-					yield Button('Export CSV', variant = 'primary', id = 'services-export-instances-button', classes = 'button')
+					with Horizontal():
+						yield Button('Export CSV', variant = 'primary', id = 'services-export-instances-button', classes = 'button')
+						yield Button('Copy CSV', variant = 'primary', id = 'services-copy-instances-button', classes = 'button')
 					yield LoadingIndicator(id = 'services-export-instances-loading-indicator', classes = 'loading-indicator')
 					yield Static('', id = 'services-export-instances-result', classes = 'result')
 
@@ -137,8 +141,9 @@ class ACMEContainerResourceServices(Container):
 			count, filename = CSE.console.doExportResource(self.resource.ri, self.exportIncludingChildResources)
 			self.exportResourceLoadingIndicator.display = False
 			self.exportResourceResult.display = True
-			self.exportResourceResult.update(f'Exported [{self._app.objectColor}]{count}[/] resource(s) to file [{self._app.objectColor}]{filename}[/]')
-	
+			self.exportResourceResult.update(n := f'Exported [{self._app.objectColor}]{count}[/] resource(s) to file [{self._app.objectColor}]{filename}[/]')
+			self._app.showNotification(n, 'Export Resources', 'information')
+
 
 		# Show the loading indicator instead of the result
 		self.exportResourceLoadingIndicator.display = True
@@ -154,15 +159,15 @@ class ACMEContainerResourceServices(Container):
 		
 	@on(Button.Pressed, '#services-export-instances-button')
 	def exportInstances(self) -> None:
-		"""	Callback to export the current resource's instances
+		"""	Callback to export the current resource's instances as CSV.
 		"""
 
 		def _exportInstances() -> None:
 			count, filename = CSE.console.doExportInstances(self.resource.ri)
 			self.exportInstancesLoadingIndicator.display = False
 			self.exportInstancesResult.display = True
-			self.exportInstancesResult.update(f"Exported [{self._app.objectColor}]{count}[/] data point(s) to file [@click=open_file('{filename}')]{filename}[/]")
-	
+			self.exportInstancesResult.update(n := f"Exported [{self._app.objectColor}]{count}[/] data point(s) to file [@click=open_file('{filename}')]{filename}[/]")
+			self._app.showNotification(n, 'Export Data Points', 'information')
 
 		# Show the loading indicator instead of the result
 		self.exportInstancesLoadingIndicator.display = True
@@ -171,3 +176,25 @@ class ACMEContainerResourceServices(Container):
 		# Execute in the background to not block the UI
 		BackgroundWorkerPool.runJob(_exportInstances)
 	
+
+	@on(Button.Pressed, '#services-copy-instances-button')
+	def copyInstances(self) -> None:
+		"""	Callback to copy the current resource's instances to the clipboard as CSV.
+		"""
+
+		def _copyInstances() -> None:
+			count, data = CSE.console.doExportInstances(self.resource.ri, asString = True)
+			self.exportInstancesLoadingIndicator.display = False
+			self.exportInstancesResult.display = True
+			pyperclip.copy(data)
+			self.exportInstancesResult.update(n := f'Copied [{self._app.objectColor}]{count}[/] data point(s) to the clipboard')
+			self._app.showNotification(n, 'Copy Data Points', 'information')
+
+
+		# Show the loading indicator instead of the result
+		self.exportInstancesLoadingIndicator.display = True
+		self.exportInstancesResult.display = False
+
+		# Execute in the background to not block the UI
+		BackgroundWorkerPool.runJob(_copyInstances)
+